+++ mdObject/src/classes/MdObject.ts
@@ -3,16 +3,12 @@ import { LocationType, UserCallFunction } from '../enums/enums'
 import { ObsTermsMap } from './ObsTermsMap';
 import { ClinicalDocument } from './ClinicalDocument';
 import { Emr } from './Emr';
-import { Users } from './Users';
-import { Patient } from './Patient';
 
 export class MdObject {
 
     private _emr = new Emr(this._window, this._document);
     private _obsTermsMap = new ObsTermsMap();
     private _clinicalDocument = new ClinicalDocument(this.emr.emrMel);
-    private _users = new Users(this.emr.emrMel);
-    private _patient = new Patient(this.obsTermsMap.weight, this.obsTermsMap.height, this.emr.emrMel, this.emr.demographics);
 
     constructor(
         public _window: any,
@@ -41,14 +37,6 @@ export class MdObject {
 
     get clinicalDocument():ClinicalDocument {
         return this._clinicalDocument;
-    }
-
-    get users():Users {
-        return this._users;
-    }
-
-    get patient():Patient {
-        return this._patient;
     }
 
     get LocationType() {